chore(core): separate unit & integration tests for webmangler/languages
