+++ packages/core/src/languages/builtin/__tests__/integration/class.test.ts
@@ -1,6 +1,6 @@
 import { expect } from "chai";
 
-import BuiltInLanguagesPlugin from "../builtin";
+import BuiltInLanguagesPlugin from "../../index";
 
 suite("Built-in Language Supports", function() {
   test("get embeds", function() {
